LMB-517: move is mandataris active and is fractie onafhankelijk to the services
